Move file system archive code into backup package
